fixing erros

+++ BearPanel/lib/models/lesson.dart
@@ -1,7 +1,7 @@
 class Lesson {
   String title;
-  int current;
-  int total;
+  double current;
+  double total;
 
   Lesson({required this.title, required this.current, required this.total});
 }
